feat: Add getProductsById function
