+++ product-service-backend/handlers/mockData.js
@@ -1,0 +1,114 @@
+const products = [
+    {
+        "id": 1,
+        "name": "iPhone 15 Pro",
+        "brand": "Apple",
+        "price": 999.99,
+        "currency": "USD",
+        "stock": 20,
+        "rating": 4.8,
+        "features": ["6.1-inch OLED display", "A17 Bionic chip", "Triple-camera system", "USB-C"],
+        "image": "https://example.com/images/iphone15pro.jpg"
+    },
+        {
+        "id": 2,
+        "name": "Samsung Galaxy S23 Ultra",
+        "brand": "Samsung",
+        "price": 1199.99,
+        "currency": "USD",
+        "stock": 15,
+        "rating": 4.7,
+        "features": ["6.8-inch AMOLED display", "Snapdragon 8 Gen 2", "200MP camera", "S Pen support"],
+        "image": "https://example.com/images/galaxys23ultra.jpg"
+        },
+    {
+    "id": 3,
+    "name": "Google Pixel 8 Pro",
+    "brand": "Google",
+    "price": 899.99,
+    "currency": "USD",
+    "stock": 25,
+    "rating": 4.6,
+    "features": ["6.7-inch LTPO OLED", "Tensor G3 chip", "50MP main camera", "Android 14"],
+    "image": "https://example.com/images/pixel8pro.jpg"
+    },
+    {
+    "id": 4,
+    "name": "OnePlus 11",
+    "brand": "OnePlus",
+    "price": 699.99,
+    "currency": "USD",
+    "stock": 30,
+    "rating": 4.5,
+    "features": ["6.7-inch AMOLED", "Snapdragon 8 Gen 2", "100W fast charging", "Hasselblad camera"],
+    "image": "https://example.com/images/oneplus11.jpg"
+    },
+    {
+    "id": 5,
+    "name": "Xiaomi 13 Pro",
+    "brand": "Xiaomi",
+    "price": 799.99,
+    "currency": "USD",
+    "stock": 18,
+    "rating": 4.4,
+    "features": ["6.73-inch OLED", "Snapdragon 8 Gen 2", "Leica cameras", "120W fast charging"],
+    "image": "https://example.com/images/xiaomi13pro.jpg"
+    },
+    {
+    "id": 6,
+    "name": "Sony Xperia 1 V",
+    "brand": "Sony",
+    "price": 1099.99,
+    "currency": "USD",
+    "stock": 10,
+    "rating": 4.3,
+    "features": ["6.5-inch 4K OLED", "Snapdragon 8 Gen 2", "Triple 12MP cameras", "Cinema Pro mode"],
+    "image": "https://example.com/images/xperia1v.jpg"
+    },
+    {
+    "id": 7,
+    "name": "Asus ROG Phone 7",
+    "brand": "Asus",
+    "price": 999.99,
+    "currency": "USD",
+    "stock": 12,
+    "rating": 4.6,
+    "features": ["6.78-inch AMOLED", "Snapdragon 8 Gen 2", "6000mAh battery", "165Hz refresh rate"],
+    "image": "https://example.com/images/rogphone7.jpg"
+    },
+    {
+    "id": 8,
+    "name": "Nothing Phone (2)",
+    "brand": "Nothing",
+    "price": 599.99,
+    "currency": "USD",
+    "stock": 22,
+    "rating": 4.4,
+    "features": ["6.7-inch OLED", "Snapdragon 8+ Gen 1", "Glyph Interface", "50MP dual cameras"],
+    "image": "https://example.com/images/nothingphone2.jpg"
+    },
+    {
+    "id": 9,
+    "name": "Motorola Edge 40 Pro",
+    "brand": "Motorola",
+    "price": 799.99,
+    "currency": "USD",
+    "stock": 17,
+    "rating": 4.3,
+    "features": ["6.67-inch OLED", "Snapdragon 8 Gen 2", "125W fast charging", "50MP camera"],
+    "image": "https://example.com/images/edge40pro.jpg"
+    },
+    {
+    "id": 10,
+    "name": "Realme GT 3",
+    "brand": "Realme",
+    "price": 699.99,
+    "currency": "USD",
+    "stock": 20,
+    "rating": 4.5,
+    "features": ["6.74-inch AMOLED", "Snapdragon 8+ Gen 1", "240W fast charging", "50MP camera"],
+    "image": "https://example.com/images/realmegt3.jpg"
+    }
+]
+
+module.exports = products;
