+++ product-service-backend/serverless.yml
@@ -18,3 +18,13 @@ functions:
       - http:
           path: products
           method: get
+  getProductsById:
+    handler: handlers/getProduct.getProduct
+    events:
+      - http:
+          path: products/{productId}
+          method: get
+          request:
+            parameters:
+              paths:
+                productId: true
